Add apic.drop_unknown_alerts config option
This allows us to skip any of the 40k+ "condition rules" that APICs can
spit out that aren't standard ACI faults. Since we are autogenerating
the full list of ACI faults into the alerts config we shouldn't have to
worry about missing any of the real ones.

Part of cisco-kusanagi/kusanagi-cloudops#122