Allow multiple ACI hosts
Add a new option called aci-hosts instead of aci-host that allows
passing multiple hosts seperated by ','. Also, fixed the consul query
and added a way for balancing the load between multiple instances. The
parameter is called aci-group-host, when the instance performs the query
it will filter by the corresponding group-host. Which means, ACI hosts
will be grouped by this key. If nothing is passed, then all the hosts
are processed i.e. single instance is deployed.